manual: more refactors
